Adding support for bill payment services including: mtn, zain, sudani, e15, etc
We simply use methods overloading to differentiate between special payment APIs and those of bill payment apis.

In order to use the bill payment api, we added types.TypeName for E15, telecos and other services. With the specific type we can set the request fields including (payeeId and paymentInfo) correctly.